Switch to typescript
