+++ README.md
@@ -1,9 +1,11 @@
 <h1 align="center">eris-commando</h1>
 
 ---------------
-`eris-commando` is a command router (framework) for the [Eris](https://abal.moe/Eris) library. The name is a play-on word of Discord.js' worst "official" framework, [discord.js-commando](https://github.com/discordjs/Commando) (Funny that the argument / argument collector code is reused here!). I initially started this in 2017 since I had minimum knowledge of Eris and their command framework is shit for the most part but still works; I made this repository and then deleted because I had other projects to do. On Nov 2018, this framework exploded with updates (0.2.0 -> 0.4.0) by myself, the initial starter in 2017. For Commando developers: I think the framework is awesome with it's builtin arguments but I hate the rest of it... I descrise that this is not a clone of the framework but the Argument/Types code were modified and used in this framework! (~~don't do a dmca attack :c~~)
+`eris-commando` is a command router (framework) for the [Eris](https://abal.moe/Eris) library. The name of the library is a play-on word that *barely* anyone uses or hates the library called `discord.js-commando`. Why did I name it `eris-commando`? Because why not.
 
 ## Installation
+> Note: **Since the library is in TypeScript, it will download the `typescript` & `@types/node` dependencies. If the bot is in JS, it will not defuse the situation, it's still usable.**
+>
 > Master Branch:
 
 ```sh
@@ -23,7 +25,7 @@ $ yarn add github:auguwu/eris-commando#dev
 ```
 
 ## Examples
-If you want examples, view the [examples/](https://github.com/auguwu/eris-commando/tree/dev/examples) directory for examples related to this framework.
+If you want examples, view the [examples](https://github.com/auguwu/eris-commando/tree/dev/examples) directory for examples related to this framework.
 
 ## Bots that use eris-commando:
 * [Noel](https://github.com/auguwu/Noel): made by August#5820 (Verified for new users)
