+++ lib/types/scheduler-info.ts
@@ -1,0 +1,4 @@
+export type SchedulerInfo = {
+    name: string;
+    interval: number;
+}
