--- src/lib/argument.js
@@ -1,1 +1,0 @@
-module.exports = class ArgumentCollector {};
