--- src/lib/interfaces/scheduler.js
@@ -1,7 +1,0 @@
-module.exports = class Scheduler {
-    constructor(info) {
-        this.name = info.name;
-        this.interval = info.interval;
-        this.fn = info.run;
-    }
-};
