--- src/lib/util/discord.js
@@ -1,11 +1,0 @@
-module.exports = class DiscordUtil {
-    /**
-     * Check if something is a function
-     * 
-     * @param {any} thing The thing
-     * @returns {boolean}
-     */
-    static isFunction(thing) {
-        return (typeof thing === 'function');
-    }
-}
