--- src/lib/util/rest.js
@@ -1,127 +1,0 @@
-const { trim } = require('./array');
-
-module.exports = class RESTClient {
-    constructor(client) { this.client = client; }
-    
-    getRole(query, guild) {
-        return new Promise((resolve, reject) => {
-            if (/^\d+$/.test(query)) {
-                const role = guild.roles.get(query);
-                if (role)
-                    return resolve(query);
-            } else if (/<@%(\d+)>$/.test(query)) {
-                const match = query.match(/^<@&(\d+)>$/);
-                const role  = guild.roles.get(match[1]);
-                if (role)
-                    return resolve(query);
-            } else {
-                const roles = guild.roles.filter(
-                    (role) => role.name.toLowerCase().includes(query.toLowerCase())
-                );
-                if (roles.length > 0)
-                    return resolve(roles[0]);
-            }
-
-            reject();
-        });
-    }
-
-    getUser(query) {
-        return new Promise((resolve, reject) => {
-            if (/^\d+$/.test(query)) {
-                const user = this.bot.users.get(query);
-                if (user) return resolve(user);
-            } else if (/^<@!?(\d+)>$/.test(query)) {
-                const match = query.match(/^<@!?(\d+)>$/);
-                const user = this.bot.users.get(match[1]);
-                if (user) return resolve(user);
-            } else if (/^(.+)#(\d{4})$/.test(query)) {
-                const match = query.match(/^(.+)#(\d{4})$/);
-                const users = this.bot.users.filter((user) => user.username === match[1] && Number(user.discriminator) === Number(match[2]));
-                if (users.length > 0) return resolve(users[0]);
-            } else {
-                const users = this.bot.users.filter((user) => user.username.toLowerCase().includes(query.toLowerCase()));
-                if (users.length > 0) return resolve(users[0]);
-            }
-            reject();
-        });
-    }
-
-    getGuild(query) {
-        return new Promise((resolve, reject) => {
-            if (/^\d+$/.test(query)) {
-                const guild = this.this.bot.guilds.get(query);
-                if (guild) return resolve(guild);
-            } else {
-                const guilds = this.this.bot.guilds.filter((guild) => guild.name.toLowerCase().includes(query.toLowerCase()));
-                if (guilds.length > 0) return resolve(guilds[0]);
-            }
-
-            reject();
-        });
-    }
-
-    getChannel(query, guild) {
-        return new Promise((resolve, reject) => {
-            if (/^\d+$/.test(query)) {
-                if (guild) {
-                    if (!guild.channels.has(query)) reject();
-                    resolve(guild.channels.get(query));
-                } else {
-                    const channel = channel in this.bot.channelGuildMap && this.bot.guilds.get(this.bot.channelGuildMap[query]).channels.get(query);
-                    if (channel) return resolve(channel);
-                }
-            } else if (/^<#(\d+)>$/.test(channel)) {
-                const match = query.match(/^<#(\d+)>$/);
-                if (guild) {
-                    if (!guild.channels.has(match[1])) reject();
-                    resolve(guild.channels.get(match[1]));
-                } else {
-                    const channel = match[1] in this.bot.channelGuildMap && this.bot.guilds.get(this.bot.channelGuildMap[match[1]]).channels.get(query);
-                    if (channel) return resolve(channel);
-                }
-            } else if (guild) {
-                const channel = guild.channels.filter((channel) => channel.name.toLowerCase().includes(query.toLowerCase()));
-                if (channel.length > 0) return resolve(channel[0]);
-            } 
-      
-            reject();
-        });
-    }
-
-    getMessage(options) {
-        return new Promise(async(resolve, reject) => {
-            try {
-                const message = await this.bot.getMessage(options.channelID, options.messageID);
-                resolve(message);
-            } catch(ex) {
-                reject('Unknown Message ID.');
-            }
-        });
-    }
-
-    getGuildEmojis(guild, len = 50) {
-        return trim(guild.emojis.map(i => `<${i.animated ? 'a' : ''}:${i.name}:${i.id}>`), len).join(', ');
-    }
-
-    getMember(query, guild, noGuessing = false) {
-        return new Promise((resolve, reject) => {
-            if (/^\d+$/.test(query)) {
-                const user = guild.members.get(query);
-                if (user) return resolve(user);
-            } else if (/^<@!?(\d+)>$/.test(query)) {
-                const match = query.match(/^<@!?(\d+)>$/);
-                const user = guild.members.get(match[1]);
-                if (user) return resolve(user);
-            } else if (/^(.+)#(\d{4})$/.test(query)) {
-                const match = query.match(/^(.+)#(\d{4})$/);
-                const users = guild.members.filter((user) => user.user.username === match[1] && Number(user.user.discriminator) === Number(match[2]));
-                if (users.length > 0) return resolve(users[0]);
-            } else if (!noGuessing) {
-                const users = guild.members.filter((user) => user.user.username.toLowerCase().includes(query.toLowerCase()));
-                if (users.length > 0) return resolve(users[0]);
-            }
-            reject();
-        });
-    }
-};
