+++ tsconfig.json
@@ -1,0 +1,24 @@
+{
+    "compilerOptions": {
+        "target": "esnext",
+        "module": "commonjs",
+        "lib": [
+            "es2015",
+            "es2016",
+            "esnext"
+        ],
+        "declaration": false,
+        "outDir": "./build",
+        "rootDir": "./lib",
+        "strict": true,
+        "noImplicitAny": false,
+        "moduleResolution": "node",
+        "types": [
+            "node"
+        ],
+        "allowSyntheticDefaultImports": true,
+        "esModuleInterop": true,
+        "experimentalDecorators": true,
+        "emitDecoratorMetadata": true
+    }
+}
